Refactored headers
